ai-services/main.py
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import List, Optional, Dict, Any
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(
    title="YojanaSaathi AI Services",
    description="AI/ML services for scheme recommendations, NLP chatbot, OCR, and fraud detection",
    version="1.0.0"
)

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Health check
@app.get("/health")
async def health_check():
    return {"status": "ok", "service": "ai-services"}

# Recommendation endpoint
@app.post("/api/recommendations")
async def get_recommendations(user_profile: Dict[str, Any]):
    """
    Get personalized scheme recommendations based on user profile
    """
    # TODO: Implement recommendation engine
    return {
        "recommendations": [],
        "message": "Recommendation engine not yet implemented"
    }

# Chatbot endpoint
@app.post("/api/chatbot")
async def chatbot(message: str, language: str = "en"):
    """
    Process chatbot messages with NLP
    """
    # TODO: Implement NLP chatbot
    return {
        "response": "Chatbot not yet implemented",
        "intent": None,
        "entities": []
    }

# OCR endpoint
@app.post("/api/ocr")
async def process_ocr(document_type: str, image_path: str):
    """
    Process document with OCR
    """
    # TODO: Implement OCR service
    return {
        "extracted_data": {},
        "message": "OCR service not yet implemented"
    }

# Fraud detection endpoint
@app.post("/api/fraud-detection")
async def detect_fraud(application_data: Dict[str, Any]):
    """
    Detect potential fraud in applications
    """
    # TODO: Implement fraud detection
    return {
        "risk_score": 0,
        "flags": [],
        "message": "Fraud detection not yet implemented"
    }

# Voice to text endpoint
@app.post("/api/voice/stt")
async def speech_to_text(audio_path: str, language: str = "en"):
    """
    Convert speech to text
    """
    # TODO: Implement speech-to-text
    return {
        "text": "",
        "confidence": 0,
        "message": "Speech-to-text not yet implemented"
    }

# Text to speech endpoint
@app.post("/api/voice/tts")
async def text_to_speech(text: str, language: str = "en"):
    """
    Convert text to speech
    """
    # TODO: Implement text-to-speech
    return {
        "audio_url": "",
        "message": "Text-to-speech not yet implemented"
    }

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)
