ai-services/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI/ML Libraries
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2

# NLP
langchain==0.0.340
sentence-transformers==2.2.2

# OCR
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78

# Database
pymongo==4.6.0
redis==5.0.1

# AWS
boto3==1.34.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
